ADC-295: Fix ad to support updation of multiple attributes in a single operation
AD updates only one attribute at a time and was solely written to enableuser, disableUser, setPassWordExpiry, etc. It was meant to update only these and so was used internally.  When we use the update externally, it updates one attribute at a time and so sends series of update command each for an attribute.